Update fiveInARow.py
added a start screen, i know it looks bad but it means that u start the game with a menu instead of dealing with a terminal, also i plan to make it look nicer later, i just wanted to test its functionality for now. The check to see if the button is clicked or not is pretty awful so if anyone wants to refactor it be my guest.